Fix: 로그인과 로그아웃 기능 분리
