+++ .idea/workspace.xml
@@ -6,12 +6,8 @@
   <component name="ChangeListManager">
     <list default="true" id="77df237f-a12e-4eb1-bc1e-6cf2e42bf405" name="Changes" comment=":tada: init project">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/KI5SKApp.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/KI5SKApp.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/model/vo/Money.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/model/vo/Money.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/user/controller/LoginController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/user/controller/LoginController.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/user/controller/UserController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/user/controller/UserController.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/user/dao/UserDAO.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/user/dao/UserDAO.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/user/dao/UserDAOImpl.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/user/dao/UserDAOImpl.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/user/model/User.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/user/model/User.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
